Add a role to install common tools
This adds a small, local role that can be used to install basic tools such as
Git, tmux, and wget. This is for tools that are easily installed via the yum or
dnf package managers. The Ansible built-in for yum will figure out which to
install.

Tools that require a more intensive build process should be handled through a
separate role for that tool, such as Python or Terraform.