Fix for Python role
Cleans up the Python role, which was defaulting to install in /tmp,
which fills up on some EC2 instances. This reworks some of the install
process, which was also using too many CPU credits. This bumps the cores
down to 1 so that it doesn't chew through them all.

This also removes the swap config/cleanup that was added to help with
memory usage, since it seems to work without it.